fixed publish handling for internal and external links
