+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/content/main.js
@@ -280,6 +280,7 @@ define([
                     this.sandbox.dom.html('li[data-id="' + this.options.language + '"] a', this.options.language);
 
                     this.sandbox.emit('sulu.header.saved', data);
+                    this.showState(!!this.data.published);
                     this.sandbox.emit('sulu.labels.success.show', 'labels.success.content-save-desc', 'labels.success');
                 }
 
@@ -724,6 +725,8 @@ define([
             this.setTemplate(this.data);
             this.showDraftLabel();
 
+            this.showState(!!this.data.published);
+
             if (this.options.content === 'permissions') {
                 this.showSaveItems('permissions');
             }
@@ -797,7 +800,6 @@ define([
             setToolbarItems.call(this, 'publish', publish);
 
             setToolbarItems.call(this, 'unpublish', !!this.data.published);
-            setToolbarItems.call(this, '')
 
             if (!!saveDraft || !!savePublish || !!publish) {
                 this.sandbox.emit('sulu.header.toolbar.item.enable', 'save', false);
@@ -1061,6 +1063,16 @@ define([
             this.sandbox.util.each(hiddenItems, function(index, hiddenItem) {
                 this.sandbox.emit('sulu.header.toolbar.item.hide', hiddenItem);
             }.bind(this));
+        },
+
+        showState: function(published) {
+            if (!!published) {
+                this.sandbox.emit('sulu.header.toolbar.item.hide', 'stateTest');
+                this.sandbox.emit('sulu.header.toolbar.item.show', 'statePublished');
+            } else {
+                this.sandbox.emit('sulu.header.toolbar.item.hide', 'statePublished');
+                this.sandbox.emit('sulu.header.toolbar.item.show', 'stateTest');
+            }
         },
 
         showDraftLabel: function() {
