+++ src/Sulu/Bundle/DocumentManagerBundle/Resources/config/core.xml
@@ -138,6 +138,7 @@
 
         <service id="sulu_document_manager.subscriber.phpcr.mixin" class="Sulu\Component\DocumentManager\Subscriber\Behavior\Mapping\MixinSubscriber">
             <argument type="service" id="sulu_document_manager.metadata_factory"/>
+            <argument type="service" id="sulu_document_manager.property_encoder"/>
             <tag name="sulu_document_manager.event_subscriber"/>
         </service>
